+++ tests/playwright/shiny/bugs/1345-render-data-frame-input/app.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from shiny import App, Inputs, reactive, render, req, ui
+from shiny import App, Inputs, reactive, render, ui
 
 app_ui = ui.page_fluid(
     ui.output_data_frame("df1"),
@@ -24,8 +24,7 @@ def server(input: Inputs):
     def cell_selection():
         cell_selection = df1.input_cell_selection()
         if cell_selection is None:
-            req(cell_selection)
-            raise ValueError("Cell selection is None")
+            return "No cells selected"
         if cell_selection["type"] != "row":
             raise ValueError(
                 f"Cell selection type is not 'row': {cell_selection['type']}"